enable bf16 for EmbeddingSpMDMBlockSize1_
